requirements.txt
fastapi
uvicorn[standard]
python-dotenv
openai
yfinance
anthropic>=0.7.0
pydantic>=2.0
httpx>=0.25.0
